fix: desktop drawer
