+++ apps/web/public/dictionaries/en-US.json
@@ -290,11 +290,6 @@
     "view": "View",
     "toggle_columns": "Toggle columns"
   },
-  "statuses": {
-    "pending": "Pending",
-    "watching": "Watching",
-    "watched": "Watched"
-  },
   "list_item_actions": {
     "use_as_cover": "Use as cover",
     "status": "Status",
@@ -719,6 +714,7 @@
   "upgrade_list_message": "Subscribe to the PRO plan to get unlimited lists.",
   "activity": "Activity",
   "watched": "Watched",
+  "watching": "Watching",
   "watchlist": "Watchlist",
   "stats": "Stats",
   "where_to_watch": "Where to watch",
