+++ apps/web/public/dictionaries/pt-BR.json
@@ -290,11 +290,6 @@
     "view": "Visualizar",
     "toggle_columns": "Alternar colunas"
   },
-  "statuses": {
-    "pending": "Pendente",
-    "watching": "Assistindo",
-    "watched": "Assistido"
-  },
   "list_item_actions": {
     "use_as_cover": "Usar como capa",
     "status": "Status",
@@ -720,6 +715,7 @@
   "activity": "Atividade",
   "watched": "Assistidos",
   "watchlist": "Assistir",
+  "watching": "Assistindo",
   "stats": "Estatísticas",
   "where_to_watch": "Onde assistir",
   "unavailable": "Indisponível",
