+++ apps/web/src/app/[lang]/[username]/_components/profile-tabs.tsx
@@ -1,14 +1,6 @@
 'use client'
 
-import {
-  Activity,
-  Award,
-  BarChartHorizontalBig,
-  Check,
-  Clock,
-  List,
-  Star,
-} from 'lucide-react'
+import { Activity, Check, Clock, List, Loader, Star } from 'lucide-react'
 import { Tabs, TabsList, TabsTrigger } from '@plotwist/ui/components/ui/tabs'
 import { useLanguage } from '@/context/language'
 import { GetUsersUsername200User } from '@/api/endpoints.schemas'
@@ -38,49 +30,56 @@ export const ProfileTabs = ({ user }: ProfileTabsProps) => {
         <TabsList>
           <TabsTrigger value="activity" asChild>
             <Link href={`/${language}/${user.username}`}>
-              <Activity className="mr-1" width={12} height={12} />
+              <Activity className="mr-1" size={12} />
               {dictionary.activity}
             </Link>
           </TabsTrigger>
 
           <TabsTrigger value="watched" asChild>
             <Link href={`/${language}/${user.username}/watched`}>
-              <Check className="mr-1" width={12} height={12} />
+              <Check className="mr-1" size={12} />
               {dictionary.watched}
+            </Link>
+          </TabsTrigger>
+
+          <TabsTrigger value="watching" asChild>
+            <Link href={`/${language}/${user.username}/watching`}>
+              <Loader className="mr-1" size={12} />
+              {dictionary.watching}
             </Link>
           </TabsTrigger>
 
           <TabsTrigger value="watchlist" asChild>
             <Link href={`/${language}/${user.username}/watchlist`}>
-              <Clock className="mr-1" width={12} height={12} />
+              <Clock className="mr-1" size={12} />
               {dictionary.watchlist}
             </Link>
           </TabsTrigger>
 
           <TabsTrigger value="lists" asChild>
             <Link href={`/${language}/${user.username}/lists`}>
-              <List className="mr-1" width={12} height={12} />
+              <List className="mr-1" size={12} />
               {dictionary.profile.lists}
             </Link>
           </TabsTrigger>
 
           <TabsTrigger value="reviews" asChild>
             <Link href={`/${language}/${user.username}/reviews`}>
-              <Star className="mr-1" width={12} height={12} />
+              <Star className="mr-1" size={12} />
               {dictionary.profile.reviews}
             </Link>
           </TabsTrigger>
 
-          <TabsTrigger value="achievements" disabled>
-            <Award className="mr-1" width={12} height={12} />
+          {/* <TabsTrigger value="achievements" disabled>
+            <Award className="mr-1" size={12} />
 
             {dictionary.profile.achievements}
           </TabsTrigger>
 
           <TabsTrigger value="achievements" disabled>
-            <BarChartHorizontalBig className="mr-1" width={12} height={12} />
+            <BarChartHorizontalBig className="mr-1" size={12} />
             {dictionary.stats}
-          </TabsTrigger>
+          </TabsTrigger> */}
         </TabsList>
       </div>
     </Tabs>
