+++ apps/web/src/app/[lang]/[username]/_components/user-items-command.tsx
@@ -11,12 +11,13 @@ import { APP_QUERY_CLIENT } from '@/context/app'
 import { useLanguage } from '@/context/language'
 import { Plus } from 'lucide-react'
 import { toast } from 'sonner'
+import { UserItemsProps } from './user-items'
 
-type WatchListCommandProps = {
+type UserItemsCommandProps = {
   items: GetUserItems200Item[]
-}
+} & Pick<UserItemsProps, 'status'>
 
-export function WatchListCommand({ items }: WatchListCommandProps) {
+export function UserItemsCommand({ items, status }: UserItemsCommandProps) {
   const add = usePostUserItem()
   const remove = useDeleteUserItemId()
 
@@ -27,13 +28,13 @@ export function WatchListCommand({ items }: WatchListCommandProps) {
       items={items}
       onAdd={(tmdbId, mediaType) =>
         add.mutate(
-          { data: { tmdbId, mediaType, status: 'WATCHLIST' } },
+          { data: { tmdbId, mediaType, status } },
           {
             onSuccess: async () => {
               await APP_QUERY_CLIENT.invalidateQueries({
                 queryKey: getGetUserItemsQueryKey({
                   language,
-                  status: 'WATCHLIST',
+                  status,
                 }),
               })
