+++ apps/web/src/app/[lang]/[username]/watching/page.tsx
@@ -1,0 +1,5 @@
+import { UserItems } from '../_components/user-items'
+
+export default async function WatchingPage() {
+  return <UserItems status="WATCHING" />
+}
