+++ apps/web/src/app/[lang]/[username]/watchlist/page.tsx
@@ -1,22 +1,5 @@
-import { Button } from '@plotwist/ui/components/ui/button'
-import { Items } from './_items'
+import { UserItems } from '../_components/user-items'
 
 export default function WatchListPage() {
-  return (
-    <div className="space-y-4">
-      <section className="grid grid-cols-3 md:grid-cols-5 gap-2">
-        <Items />
-      </section>
-
-      <div className="flex justify-between">
-        <Button size="sm" variant="outline">
-          Newest
-        </Button>
-
-        <Button size="sm" variant="outline">
-          Oldest
-        </Button>
-      </div>
-    </div>
-  )
+  return <UserItems status="WATCHLIST" />
 }
