+++ apps/web/src/app/[lang]/tv-series/[id]/_components/tv-serie-details.tsx
@@ -6,6 +6,7 @@ import { Language } from '@/types/languages'
 import { tmdb } from '@/services/tmdb'
 import { TvSerieInfos } from './tv-serie-infos'
 import { TvSerieTabs } from './tv-serie-tabs'
+import { ItemDrawer } from '@/components/item-drawer'
 
 type TvSerieDetailsProps = {
   id: number
@@ -16,9 +17,13 @@ export const TvSerieDetails = async ({ id, language }: TvSerieDetailsProps) => {
   const tvSerie = await tmdb.tv.details(id, language)
 
   return (
-    <div className="mx-auto max-w-6xl">
+    <div className="mx-auto max-w-6xl relative">
+      <ItemDrawer
+        mediaType="TV_SHOW"
+        tmdbId={tvSerie.id}
+        title={`${tvSerie.name}`}
+      />
       <Banner url={tmdbImage(tvSerie.backdrop_path)} />
-
       <section className="mx-auto my-8 max-w-4xl space-y-6">
         <TvSerieInfos language={language} tvSerie={tvSerie} />
         <TvSerieTabs language={language} tvSerie={tvSerie} />
